Create ReceiveDetail component
