+++ src/pages/assets/assets.tsx
@@ -1,13 +1,11 @@
 import React, { useState, useEffect, useRef } from 'react';
-import QRCode from 'qrcode.react';
 import { useRecoilValue } from 'recoil';
 import numeral from 'numeral';
 import { useTranslation } from 'react-i18next';
 import './assets.less';
 import 'antd/dist/antd.css';
-import { CopyToClipboard } from 'react-copy-to-clipboard';
-import { Layout, Button, Table, Avatar, notification } from 'antd';
-import { CopyOutlined, ArrowLeftOutlined } from '@ant-design/icons';
+import { Layout, Table, Avatar } from 'antd';
+import { ArrowLeftOutlined } from '@ant-design/icons';
 import {
   sessionState,
   marketState,
@@ -15,11 +13,12 @@ import {
   walletAllAssetsState,
 } from '../../recoil/atom';
 import { Session } from '../../models/Session';
-import { getAssetBalancePrice, UserAsset, UserAssetType } from '../../models/UserAsset';
+import { getAssetBalancePrice, UserAsset } from '../../models/UserAsset';
 import { getUIDynamicAmount } from '../../utils/NumberUtils';
-import { LEDGER_WALLET_TYPE, createLedgerDevice } from '../../service/LedgerService';
+// import { LEDGER_WALLET_TYPE, createLedgerDevice } from '../../service/LedgerService';
 import { AnalyticsService } from '../../service/analytics/AnalyticsService';
 // import logoCro from '../../assets/AssetLogo/cro.png';
+import ReceiveDetail from './components/ReceiveDetail';
 
 const { Header, Content, Footer } = Layout;
 
@@ -28,7 +27,7 @@ const AssetsPage = () => {
   const userAsset = useRecoilValue(walletAssetState);
   const walletAllAssets = useRecoilValue(walletAllAssetsState);
   const marketData = useRecoilValue(marketState);
-  const [isLedger, setIsLedger] = useState(false);
+  // const [isLedger, setIsLedger] = useState(false);
   const [currentAsset, setCurrentAsset] = useState<UserAsset>();
   const [isAssetVisible, setIsAssetVisible] = useState(false);
   const didMountRef = useRef(false);
@@ -37,43 +36,11 @@ const AssetsPage = () => {
   const [t] = useTranslation();
 
   useEffect(() => {
-    const { walletType } = session.wallet;
-    setIsLedger(LEDGER_WALLET_TYPE === walletType);
     if (!didMountRef.current) {
       didMountRef.current = true;
       analyticsService.logPage('Assets');
     }
   });
-
-  const clickCheckLedger = async () => {
-    try {
-      const { addressIndex, walletType, config } = session.wallet;
-      const addressprefix = config.network.addressPrefix;
-      if (LEDGER_WALLET_TYPE === walletType) {
-        const device = createLedgerDevice();
-        await device.getAddress(addressIndex, addressprefix, true);
-      }
-    } catch (e) {
-      notification.error({
-        message: t('receive.notification.ledgerConnect.message'),
-        description: t('receive.notification.ledgerConnect.description'),
-        placement: 'topRight',
-        duration: 3,
-      });
-    }
-  };
-
-  const onCopyClick = () => {
-    setTimeout(() => {
-      notification.success({
-        message: t('receive.notification.addressCopy.message'),
-        description: t('receive.notification.addressCopy.description'),
-        placement: 'topRight',
-        duration: 2,
-        key: 'copy',
-      });
-    }, 100);
-  };
 
   const assetIcon = asset => {
     const { icon_url, symbol } = asset;
@@ -83,23 +50,6 @@ const AssetsPage = () => {
     ) : (
       <Avatar>{symbol[0].toUpperCase()}</Avatar>
     );
-  };
-
-  const assetAddress = (asset, _session) => {
-    const { assetType, address } = asset;
-    // For IBC assets
-    const { wallet } = _session;
-
-    switch (assetType) {
-      case UserAssetType.TENDERMINT:
-        return address;
-      case UserAssetType.EVM:
-        return address;
-      case UserAssetType.IBC:
-        return wallet.address;
-      default:
-        return wallet.address;
-    }
   };
 
   const AssetColumns = [
@@ -188,27 +138,7 @@ const AssetsPage = () => {
                       Asset List
                     </div>
                   </a>
-                  <div className="title">
-                    {assetIcon(currentAsset)}
-                    {currentAsset?.name} ({currentAsset?.symbol})
-                  </div>
-                  <div className="address">
-                    <QRCode value={assetAddress(currentAsset, session)} size={180} />
-                    <div className="name">{session.wallet.name}</div>
-                  </div>
-                  <CopyToClipboard text={assetAddress(currentAsset, session)}>
-                    <div className="copy" onClick={onCopyClick}>
-                      {assetAddress(currentAsset, session)}
-                      <CopyOutlined />
-                    </div>
-                  </CopyToClipboard>
-                  {isLedger && (
-                    <div className="ledger">
-                      <Button type="primary" onClick={clickCheckLedger}>
-                        {t('receive.button')}
-                      </Button>
-                    </div>
-                  )}
+                  <ReceiveDetail currentAsset={currentAsset} session={session} />
                 </Content>
               </Layout>
             ) : (
